pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "svachalan"
version = "0.1.0"
description = "Deterministic browser automation runner for Chromium workflows."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
  "httpx>=0.28.1",
  "pydantic>=2.11.1",
  "PyYAML>=6.0.2",
  "websockets>=15.0.1",
]

[project.scripts]
svachalan = "svachalan.cli.main:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "ruff>=0.13.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
